Many Changes -112 Commits (#756)
Changes:
- Fix mistakes in the default config.
- Added `/ce bottle` command.
- Added `/ce give` command.
- Fix item names being set to "" when there is no name set.
- Allow for " " to be set as a name on items.
- Add back drop toggle to veinMiner and blast.
- Make `/ce remove` only suggest enchants that are on the item.
- Make some methods pass a copy instead of the pointer to avoid overlap between handling.
- Fix players in spectator mode with wings not being able to fly.
- Fix telepathy not being able to work on stairs.
- Fix random enchantment level.
- Better isInt() check.
- TempFix factions errors caused by forks.
- Add an option to pick if TransmogScroll should add spaces between sections.
- Added config option to change the sound played for each enchant being added to an item.
- Fixed GKitz timer not displaying time in the correct format.
- Make ally mobs no longer target each other.
- Make the max for random enchants inclusive.
- Make random enchant levels not apply enchants that roll a 0.
- Fix Aura enchants not working.
- #774 
  - Added Per-Item Max Enchant
  - Slot Crystal
- #775
  - Make ally mobs not attack mcmmo party members.
  - Adds a cap on the number of times Aura enchantments can activate to reduce server load.
- #776 
  - Change it from individual checks over to using a single method.
- Changed Enchants over to each having their own timer.
  - Each enchant can only be used by each player once per second.
  - New permissions `crazyenchantments.%s.deny`
- Removed enchant timer for now.
- Started to change all armor enchantment checks to new methods.
- Removed silk spawner support. (if they handle it properly, enchants should still work with it)
- Properly create a block drop item event for every block break.
- #748 done by @Mowstyl 

Internal:
- Mass cleanup.
- Better methods.
- Less clutter.
- Better performance.

Bug fixes:
- All found along the way.

Misc
- Can't remember when, but `crazyenchantments.blast.use` was removed.

---------

+++ paper/src/main/java/com/badbones69/crazyenchantments/paper/api/economy/CurrencyAPI.java
@@ -156,19 +156,19 @@ public class CurrencyAPI {
             experience = (int) Math.ceil(Math.pow(level, 2) + (6 * level));
             int requiredExperience = 2 * level + 7;
             double currentExp = Double.parseDouble(Float.toString(player.getExp()));
-            experience += Math.ceil(currentExp * requiredExperience);
+            experience += (int) Math.ceil(currentExp * requiredExperience);
             return experience;
         } else if (level > 15 && level <= 30) {
             experience = (int) Math.ceil((2.5 * Math.pow(level, 2) - (40.5 * level) + 360));
             int requiredExperience = 5 * level - 38;
             double currentExp = Double.parseDouble(Float.toString(player.getExp()));
-            experience += Math.ceil(currentExp * requiredExperience);
+            experience += (int) Math.ceil(currentExp * requiredExperience);
             return experience;
         } else {
             experience = (int) Math.ceil((4.5 * Math.pow(level, 2) - (162.5 * level) + 2220));
             int requiredExperience = 9 * level - 158;
             double currentExp = Double.parseDouble(Float.toString(player.getExp()));
-            experience += Math.ceil(currentExp * requiredExperience);
+            experience += (int) Math.ceil(currentExp * requiredExperience);
             return experience;
         }
     }
